[#1] Adjust to matchers API changes in 2.0.22+

+++ src/main/java/info/solidsoft/mockito/java8/api/WithMockito.java
@@ -285,9 +285,9 @@ public interface WithMockito extends WithMatchers {
     }
 
     /**
-     * Delegates call to {@link Mockito#after(int)}.
+     * Delegates call to {@link Mockito#after(long)}.
      */
-    default VerificationAfterDelay after(int millis) {
+    default VerificationAfterDelay after(long millis) {
         return Mockito.after(millis);
     }
